requirements.txt
# Core libraries
requests==2.31.0
pandas==2.1.4
psycopg2-binary==2.9.9
